refactor: update rtcp protocol structure;
