--- rtcp/src/protocol/receiver_report.rs
@@ -1,4 +1,0 @@
-
-pub struct ReceiverReport {
-    
-}
